Moved flags series into a subfolder.
